// src/app/interfaces/disco.ts
export interface Disco {
  [key: string]: number | string | null;
  axle: string | null;
  style: string | null;
  diameter: number | null;
  height: number | null;
  thicknessNew: number | null;
  thicknessMin: number | null;
  pcd: number | null;
  holes: number | null;
  centerbore: number | null;



}